make the details page a server component
