+++ components/MovieViewTracker.tsx
@@ -3,10 +3,8 @@
 import { useEffect } from "react";
 import { useAuth } from "./Providers";
 
-export default function DetailsClient({ movieId }: { movieId: number }) {
+export default function MovieViewTracker({ movieId }: { movieId: number }) {
   const { user } = useAuth();
-  // console.log("user in details client", user);
-  // console.log("movieId in details client", movieId);
 
   useEffect(() => {
     if (user) {
